requirements.txt
fastapi==0.110.2
ipython==8.12.3
langchain==0.1.16
langchain_core==0.1.46
langchain_google_genai==1.0.3
pydantic==2.7.1
python-dotenv==1.0.1
starlette==0.37.2
uvicorn==0.29.0
